+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The version numbers are structured like `GENERATION.MAJOR.MINOR.PATCH`:
 - Add service definition `Heptacom\HeptaConnect\Core\Portal\File\Filesystem\Contract\FilesystemFactoryInterface` for class `\Heptacom\HeptaConnect\Core\Portal\File\Filesystem\FilesystemFactory`
 - Add dependency `Heptacom\HeptaConnect\Core\Portal\File\Filesystem\Contract\FilesystemFactoryInterface` to service `Heptacom\HeptaConnect\Core\Portal\Contract\PortalStackServiceContainerBuilderInterface`
 - Add command `heptaconnect:emit` to emit one or more entities
+- Add composer suggestion `psy/psysh` for an interactive read–eval–print loop in the scope of a portal-node
+- Add command `heptaconnect:repl` for an interactive read–eval–print loop in the scope of a portal-node
 
 ### Changed